demo

+++ test/demo.py
@@ -102,7 +102,7 @@ def train(clf, data, device):
 
 def test(clf, data, device):
     clf.eval()
-    preds = torch.zeros(len(data), dtype=data.targets.dtype)
+    preds = torch.zeros(len(data), dtype=data.targets[0].dtype)
     loader = DataLoader(data, shuffle=False, **PARAMS['test_args'])
     with torch.no_grad():
         idx = 0
